# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Backend do Terminal Rodoviário",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "https://terminal-rodoviario-backend.herokuapp.com"
    },
    {
      "url": "http://localhost:8000"
    }
  ],
...
  "paths": {
...
    "/viagem": {
      "get": {
        "tags": ["Viagem"],
        "summary": "Retorna uma lista com todas as viagens ativas",
        "operationId": "fetchViagens",
        "responses": {
          "200": {
            "description": "Caso a busca seja bem sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Viagem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/create": {
      "post": {
        "tags": ["Viagem"],
        "summary": "Cria uma nova viagem",
        "operationId": "createViagem",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "viacaoId": {
                    "type": "string"
                  },
                  "destino": {
                    "type": "string"
                  },
                  "data": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "dataAntiga": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "horario": {
                    "type": "string"
                  },
                  "horarioAntigo": {
                    "type": "string"
                  },
                  "valor": {
                    "type": "number"
                  },
                  "valorAntigo": {
                    "type": "number"
                  },
                  "imagem": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Caso a busca seja bem sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Viagem"
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/scheduled": {
      "get": {
        "tags": ["Viagem"],
        "summary": "Retorna uma lista com todas as viagens marcadas",
        "operationId": "fetchScheduledTrip",
        "responses": {
          "200": {
            "description": "Caso a busca seja bem sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Viagem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Caso o id definido no token de autorização não pertença a nenhum usuário ativo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "O usuário não existe."
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/{viagemId}/schedule": {
      "post": {
        "tags": ["Viagem"],
        "summary": "Marca uma viagem",
        "operationId": "scheduleTrip",
        "parameters": [
          {
            "name": "viagemId",
            "in": "path",
            "description": "Id da viação",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caso a viagem seja marcada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viagem marcada com sucesso."
                }
              }
            }
          },
          "400": {
            "description": "Caso o id definido no token de autorização não pertença a nenhum usuário ativo ou caso o viagemId não pertença a uma viagem ativa",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "O usuário não existe": {
                    "value": {
                      "message": "O usuário não existe.."
                    }
                  },
                  "A viagem não existe": {
                    "value": {
                      "message": "A viagem não existe."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/{viagemId}/unschedule": {
      "post": {
        "tags": ["Viagem"],
        "summary": "Desmarca uma viagem",
        "operationId": "unscheduleTrip",
        "parameters": [
          {
            "name": "viagemId",
            "in": "path",
            "description": "Id da viação",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caso a viagem seja desmarcada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viagem desmarcada com sucesso."
                }
              }
            }
          },
          "400": {
            "description": "Caso o id definido no token de autorização não pertença a nenhum usuário ativo ou caso o viagemId não pertença a uma viagem ativa",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "O usuário não existe": {
                    "value": {
                      "message": "O usuário não existe.."
                    }
                  },
                  "A viagem não existe": {
                    "value": {
                      "message": "A viagem não existe."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/find": {
      "get": {
        "tags": ["Viagem"],
        "summary": "Retorna uma lista de viagens com base no filtro passado",
        "operationId": "findViagensById",
        "parameters": [
          {
            "in": "query",
            "name": "destino",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "String que servirá como base para a filtragem"
          }
        ],
        "responses": {
          "200": {
            "description": "Caso a busca seja bem sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Viagem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/viagem/delete/{viagemId}": {
      "delete": {
        "tags": ["Viagem"],
        "summary": "Deleta uma viagem com base no id passado",
        "operationId": "deleteViagem",
        "parameters": [
          {
            "name": "viagemId",
            "in": "path",
            "description": "Id da viagem",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caso a viagem seja excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viagem deletada com sucesso."
                }
              }
            }
          },
          "400": {
            "description": "Caso o id passado não corresponda a uma viagem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Viagem não existe."
                }
              }
            }
          },
          "401": {
            "description": "Caso não tenha sido enviado um token ou caso ele esteja malformatado, expirado ou inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    },
                    {
                      "$ref": "#/components/schemas/Mensagem"
                    }
                  ]
                },
                "examples": {
                  "Token não enviado": {
                    "value": {
                      "message": "Nenhum token foi enviado."
                    }
                  },
                  "Token malformatado, expirado ou inválido": {
                    "value": {
                      "message": "Token malformatado, expirado ou inválido."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Caso ocorra algum erro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "example": {
                  "message": "Algo inesperado aconteceu."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Viagem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "viacaoId": {
            "type": "string"
          },
          "destino": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "dataAntinga": {
            "type": "string",
            "format": "date-time"
          },
          "horario": {
            "type": "string"
          },
          "horarioAntigo": {
            "type": "string"
          },
          "valor": {
            "type": "number"
          },
          "valorAntigo": {
            "type": "number"
          },
          "imagemUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Mensagem": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}